gym_management/requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.22.0
mysqlclient==2.2.4
packaging==24.1
python-decouple==3.8
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
marshmallow-sqlalchemy==1.1.0
mypy==1.11.2
mypy-extensions==1.0.0
sqlalchemy-stubs==0.4